+++ test-project/src/rustfmt-default.rs
@@ -1,0 +1,5 @@
+// this file ensures that by default rustfmt is invoked without parameters
+// and formats with its default (--edition=2015 as of today)
+
+// With --edition=2024 this import will be reordered
+use std::{i16, i8};